+++ src/main/java/de/justinharder/base/persistence/JpaRepository.java
@@ -30,13 +30,13 @@ public abstract class JpaRepository<T extends Entitaet> implements Repository<T>
 	}
 
 	@Override
-	public List<T> findeAlle(@NonNull Long page, @NonNull Long pageSize)
+	public List<T> findeAlle(@NonNull Integer page, @NonNull Integer pageSize)
 	{
 		entityManager.clear();
 		var criteriaQuery = entityManager.getCriteriaBuilder().createQuery(entitaet);
 		return entityManager.createQuery(criteriaQuery.select(criteriaQuery.from(entitaet)))
-			.setFirstResult(Math.toIntExact((page - 1) * pageSize))
-			.setMaxResults(Math.toIntExact(pageSize))
+			.setFirstResult((page - 1) * pageSize)
+			.setMaxResults(pageSize)
 			.getResultList();
 	}
